@@ -1,11 +1,16 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import "semantic-ui-css/semantic.min.css";
-import "./index.css";
+import { createRoot } from "react-dom/client";
+
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+import "semantic-ui-css/semantic.min.css";
+import "./index.css";
+
+const container = document.getElementById("root");
+const root = createRoot(container);
+
+root.render(<App />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
